[REFACTOR] carNames로 명칭변경
