+++ src/Const.js
@@ -9,7 +9,7 @@ const ERROR_MESSAGE = {
 };
 
 const INPUT_MESSAGE = {
-  raceCarNames:
+  carNames:
     "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
   raceCount: "시도할 횟수는 몇 회인가요?\n",
 };
