+++ src/ui/Output.js
@@ -1,9 +1,4 @@
 class Output {
-  printResult(cars) {
-    console.log("실행 결과\n");
-    this.printCarPosition(cars);
-  }
-
   printCarPosition(car) {
     let positionStick = `${car.name} : `;
     for (let i = 0; i < car.position; i++) {
